Convert DeletionService to use IScopeProvider (#3721)
Since the DeletionService is used by the `$bulk-delete` operation, the scope of `IFhirDataStore` and `ISearchService` need to be managed manually - not by the root container. Using `IScopeProvider` enables us to manage and dispose correctly - fixing the bug of leaking memory due to growth of services tracked by the root container.